pyproject.toml
[tool.poetry]
name = "otus-hw"
version = "0.1.0"
description = ""
authors = ["helenvr"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.4"
loguru = "^0.7.3"
pyyaml = "^6.0.2"
ujson = "^5.10.0"
pytest = "^8.3.4"
black = "^25.1.0"
asyncpg = "^0.30.0"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
